+++ src/main_api.h
@@ -20,15 +20,15 @@ struct SetupPacket {
   char connectionHash[4000];
 };
 struct CreatePacket {
-  short id;
+  int32_t id;
   objid sceneId;
   char serialobj[3000];
 };
 struct DeletePacket {
-  short id;
+  int32_t id;
 };
 struct UpdatePacket {
-  short id;
+  int32_t id;
   Properties properties;
 };
 
@@ -55,59 +55,59 @@ NetworkPacket toNetworkPacket(UdpPacket& packet);
 // These functions represent the functionality that individual language bindings use, but should not be coupled to any language in particular. 
 // Should trim down when it becomes clear what the core api should.
 
-void setActiveCamera(short cameraId);
+void setActiveCamera(int32_t cameraId);
 void nextCamera();
 void moveCamera(glm::vec3 offset);  
 void rotateCamera(float xoffset, float yoffset);
 
-void applyImpulse(short index, glm::vec3 impulse);
-void applyImpulseRel(short index, glm::vec3 impulse);
-void clearImpulse(short index);
+void applyImpulse(int32_t index, glm::vec3 impulse);
+void applyImpulseRel(int32_t index, glm::vec3 impulse);
+void clearImpulse(int32_t index);
 
-void loadScriptFromWorld(std::string script, short id);
-short loadScene(std::string sceneFile);
-short loadSceneObj(std::string sceneFile, short sceneId);
-short loadSceneData(std::string sceneData, objid sceneId);
-void unloadScene(short sceneId);
+void loadScriptFromWorld(std::string script, int32_t id);
+int32_t loadScene(std::string sceneFile);
+int32_t loadSceneObj(std::string sceneFile, int32_t sceneId);
+int32_t loadSceneData(std::string sceneData, objid sceneId);
+void unloadScene(int32_t sceneId);
 void unloadAllScenes();
 void saveScene(bool includeIds);
-std::vector<short> listScenes();
-void sendLoadScene(short sceneId);
+std::vector<int32_t> listScenes();
+void sendLoadScene(int32_t sceneId);
 
 void onObjectEnter(const btCollisionObject* obj1, const btCollisionObject* obj2, glm::vec3 contactPos);
 void onObjectLeave(const btCollisionObject* obj1, const btCollisionObject* obj2);
 
 std::optional<objid> getGameObjectByName(std::string name);
-std::vector<short> getObjectsByType(std::string type);
-std::string getGameObjectName(short index);
-std::map<std::string, std::string> getGameObjectAttr(short id);
-void setGameObjectAttr(short id, std::map<std::string, std::string> attr);
+std::vector<int32_t> getObjectsByType(std::string type);
+std::string getGameObjectName(int32_t index);
+std::map<std::string, std::string> getGameObjectAttr(int32_t id);
+void setGameObjectAttr(int32_t id, std::map<std::string, std::string> attr);
 
-glm::vec3 getGameObjectPosition(short index, bool isWorld);
-void setGameObjectPosition(short index, glm::vec3 pos);
-void setGameObjectPositionRelative(short index, float x, float y, float z, bool xzPlaneOnly);
+glm::vec3 getGameObjectPosition(int32_t index, bool isWorld);
+void setGameObjectPosition(int32_t index, glm::vec3 pos);
+void setGameObjectPositionRelative(int32_t index, float x, float y, float z, bool xzPlaneOnly);
 
-glm::vec3 getGameObjectScale(short index);
-void setGameObjectScale(short index, glm::vec3 scale);
+glm::vec3 getGameObjectScale(int32_t index);
+void setGameObjectScale(int32_t index, glm::vec3 scale);
 
-glm::quat getGameObjectRotation(short index, bool isWorld);
-void setGameObjectRotation(short index, glm::quat rotation);
+glm::quat getGameObjectRotation(int32_t index, bool isWorld);
+void setGameObjectRotation(int32_t index, glm::quat rotation);
 void setSelectionMode(bool enabled);
 
-short makeObject(std::string serializedobj, objid id, bool useObjId, objid sceneId = -1, bool useSceneId = false);
+int32_t makeObject(std::string serializedobj, objid id, bool useObjId, objid sceneId = -1, bool useSceneId = false);
 objid makeObjectAttr(std::string name, std::map<std::string, std::string> stringAttributes, std::map<std::string, double> numAttributes, std::map<std::string, glm::vec3> vecAttributes);
 
-void removeObjectById(short id);
+void removeObjectById(int32_t id);
 
 void drawText(std::string word, float left, float top, unsigned int fontSize);
 
 struct AnimationState {
-  std::map<short, TimePlayback> playbacks;
+  std::map<int32_t, TimePlayback> playbacks;
 };
-std::vector<std::string> listAnimations(short id);
-void playAnimation(short id, std::string animationToPlay);
+std::vector<std::string> listAnimations(int32_t id);
+void playAnimation(int32_t id, std::string animationToPlay);
 
-std::vector<std::string> listSounds(short id);
+std::vector<std::string> listSounds(int32_t id);
 void playSound(std::string sound);
 
 std::vector<std::string> listModels();
@@ -115,8 +115,8 @@ std::vector<std::string> listModels();
 void sendEventMessage(std::string message);
 void sendNotifyMessage(std::string message);
 
-void attachToRail(short id, std::string rail);
-void unattachFromRail(short id);
+void attachToRail(int32_t id, std::string rail);
+void unattachFromRail(int32_t id);
 double timeSeconds();
 
 void sendDataUdp(std::string data);
